requirements.txt
fastapi>=0.115.5
uvicorn[standard]>=0.32.1
httpx>=0.28.1
python-dotenv>=1.0.1
pydantic>=2.10.3
pydantic-settings>=2.6.1
